requirements.txt
Django~=4.2.13
pytz~=2024.1
djangorestframework~=3.15.1
pymongo~=4.7.3
python-dotenv~=1.0.1